LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.letterSpacing.measure-expected.txt

2d.text.drawing.style.letterSpacing.measure
Testing letter spacing with different length units
Actual output:

FAIL Testing letter spacing with different length units assert_approx_equals: letter spacing doesn't work. expected 9.853906250000001 +/- 0.2 but got 18.50390625
